requirements.txt
# Fuel Cost Route Optimizer - Requirements File
# Latest Updated: December 2025 (Fixed & Verified)
# Python Version: 3.8+

# Core Django Framework
Django==4.2.9
DjangoRestFramework==3.14.0
django-cors-headers==4.3.1

django-filter==25.2

# Database & ORM
sqlparse==0.4.4
# psycopg2-binary==2.9.9  # PostgreSQL adapter (optional for production)

# API & Data Processing
requests==2.31.0
pytz==2023.3
python-decouple==3.8

# Utilities
wsgiref==0.1.2
charset-normalizer==3.3.2

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# API Documentation
drf-spectacular==0.26.5

# Environment & Configuration
python-dotenv==1.0.0

# Caching
redis==5.0.1

# Testing (Optional)
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality (Optional)
black==23.12.0
flake8==6.1.0
isort==5.13.2
pylint==3.0.3

# Security
cryptography==41.0.7
PyJWT==2.8.1

# Monitoring & Logging
django-extensions==3.2.3

# Performance
django-cache-url==3.4.4
dj-database-url==2.1.0